+++ fltk/src/image.rs
@@ -266,6 +266,14 @@ impl SvgImage {
             }
         }
     }
+
+    /// Rasterize an SvgImage
+    pub fn normalize(&mut self) {
+        assert!(!self.was_deleted());
+        unsafe {
+            Fl_SVG_Image_normalize(self.inner)
+        }
+    }
 }
 
 /// Creates a struct holding a BMP image